Pull common pytest fixtures into conftest.py
